move KeyboardAvoidingView to components/misc folder
